+++ rustica-agent/src/rustica/mod.rs
@@ -2,7 +2,7 @@ pub mod cert;
 pub mod error;
 pub mod key;
 pub mod x509;
-pub mod signers;
+pub mod signerkeys;
 
 use std::ops::Deref;
 use std::time::Duration;
