+++ tests/test_configs/rustica_local_file_alt.toml
@@ -104,3 +104,7 @@ client_certificate_authority_common_name = "RusticaAccess"
 
 [authorization."database"]
 path = "examples/example.db"
+
+[authorized_signer_keys]
+cache_validity_length.secs = 900
+cache_validity_length.nanos = 0
